# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Memorial API - Página de Recuerdos
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Persona {
  id: number;
  nombre: string;
  fechaNacimiento?: string | null;
  fechaFallecimiento?: string | null;
  biografia?: string | null;
  fotoPrincipal?: string | null;
  totalVelas: number;
  totalRecuerdos: number;
  createdAt: string;
}

export interface CreatePersonaBody {
  nombre: string;
  fechaNacimiento?: string | null;
  fechaFallecimiento?: string | null;
  biografia?: string | null;
  fotoPrincipal?: string | null;
}

export interface Vela {
  id: number;
  personaId?: number | null;
  nombreRecordado: string;
  nombreAutor: string;
  mensaje: string;
  createdAt: string;
  tiempoTranscurrido: string;
}

export interface CreateVelaBody {
  personaId?: number | null;
  nombreRecordado: string;
  nombreAutor: string;
  mensaje: string;
}

export interface VelasResponse {
  data: Vela[];
  total: number;
}

export interface Recuerdo {
  id: number;
  personaId?: number | null;
  nombreAutor: string;
  persona?: string | null;
  mensaje: string;
  fotoUrl?: string | null;
  createdAt: string;
  tiempoTranscurrido: string;
}

export interface CreateRecuerdoBody {
  personaId?: number | null;
  nombreAutor: string;
  persona?: string | null;
  mensaje: string;
  fotoUrl?: string | null;
}

export interface RecuerdosResponse {
  data: Recuerdo[];
  total: number;
}

export interface Testimonio {
  id: number;
  nombreAutor: string;
  texto: string;
  inicial: string;
  createdAt: string;
  tiempoTranscurrido: string;
}

export interface CreateTestimonioBody {
  nombreAutor: string;
  texto: string;
}

export interface Stats {
  totalPersonas: number;
  totalVelas: number;
  totalRecuerdos: number;
  totalTestimonios: number;
  velasHoy: number;
  recuerdosRecientes: Recuerdo[];
  velasRecientes: Vela[];
}

export type ListVelasParams = {
personaId?: number;
limit?: number;
offset?: number;
};

export type ListRecuerdosParams = {
personaId?: number;
limit?: number;
offset?: number;
};

export type ListTestimoniosParams = {
limit?: number;
offset?: number;
};