# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Memorial API - Página de Recuerdos
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Listar personas del memorial
 */
export const ListPersonasResponseItem = zod.object({
  "id": zod.number(),
  "nombre": zod.string(),
  "fechaNacimiento": zod.string().nullish(),
  "fechaFallecimiento": zod.string().nullish(),
  "biografia": zod.string().nullish(),
  "fotoPrincipal": zod.string().nullish(),
  "totalVelas": zod.number(),
  "totalRecuerdos": zod.number(),
  "createdAt": zod.string()
})
export const ListPersonasResponse = zod.array(ListPersonasResponseItem)


/**
 * @summary Agregar una persona al memorial
 */
export const CreatePersonaBody = zod.object({
  "nombre": zod.string(),
  "fechaNacimiento": zod.string().nullish(),
  "fechaFallecimiento": zod.string().nullish(),
  "biografia": zod.string().nullish(),
  "fotoPrincipal": zod.string().nullish()
})


/**
 * @summary Obtener una persona
 */
export const GetPersonaParams = zod.object({
  "id": zod.coerce.number()
})

export const GetPersonaResponse = zod.object({
  "id": zod.number(),
  "nombre": zod.string(),
  "fechaNacimiento": zod.string().nullish(),
  "fechaFallecimiento": zod.string().nullish(),
  "biografia": zod.string().nullish(),
  "fotoPrincipal": zod.string().nullish(),
  "totalVelas": zod.number(),
  "totalRecuerdos": zod.number(),
  "createdAt": zod.string()
})


/**
 * @summary Listar velas encendidas
 */
export const listVelasQueryLimitDefault = 20;
export const listVelasQueryOffsetDefault = 0;

export const ListVelasQueryParams = zod.object({
  "personaId": zod.coerce.number().optional(),
  "limit": zod.coerce.number().default(listVelasQueryLimitDefault),
  "offset": zod.coerce.number().default(listVelasQueryOffsetDefault)
})

export const ListVelasResponse = zod.object({
  "data": zod.array(zod.object({
  "id": zod.number(),
  "personaId": zod.number().nullish(),
  "nombreRecordado": zod.string(),
  "nombreAutor": zod.string(),
  "mensaje": zod.string(),
  "createdAt": zod.string(),
  "tiempoTranscurrido": zod.string()
})),
  "total": zod.number()
})


/**
 * @summary Encender una vela con mensaje
 */
export const CreateVelaBody = zod.object({
  "personaId": zod.number().nullish(),
  "nombreRecordado": zod.string(),
  "nombreAutor": zod.string(),
  "mensaje": zod.string()
})


/**
 * @summary Listar recuerdos
 */
export const listRecuerdosQueryLimitDefault = 20;
export const listRecuerdosQueryOffsetDefault = 0;

export const ListRecuerdosQueryParams = zod.object({
  "personaId": zod.coerce.number().optional(),
  "limit": zod.coerce.number().default(listRecuerdosQueryLimitDefault),
  "offset": zod.coerce.number().default(listRecuerdosQueryOffsetDefault)
})

export const ListRecuerdosResponse = zod.object({
  "data": zod.array(zod.object({
  "id": zod.number(),
  "personaId": zod.number().nullish(),
  "nombreAutor": zod.string(),
  "persona": zod.string().nullish(),
  "mensaje": zod.string(),
  "fotoUrl": zod.string().nullish(),
  "createdAt": zod.string(),
  "tiempoTranscurrido": zod.string()
})),
  "total": zod.number()
})


/**
 * @summary Compartir un recuerdo
 */
export const CreateRecuerdoBody = zod.object({
  "personaId": zod.number().nullish(),
  "nombreAutor": zod.string(),
  "persona": zod.string().nullish(),
  "mensaje": zod.string(),
  "fotoUrl": zod.string().nullish()
})


/**
 * @summary Listar testimonios
 */
export const listTestimoniosQueryLimitDefault = 20;
export const listTestimoniosQueryOffsetDefault = 0;

export const ListTestimoniosQueryParams = zod.object({
  "limit": zod.coerce.number().default(listTestimoniosQueryLimitDefault),
  "offset": zod.coerce.number().default(listTestimoniosQueryOffsetDefault)
})

export const ListTestimoniosResponseItem = zod.object({
  "id": zod.number(),
  "nombreAutor": zod.string(),
  "texto": zod.string(),
  "inicial": zod.string(),
  "createdAt": zod.string(),
  "tiempoTranscurrido": zod.string()
})
export const ListTestimoniosResponse = zod.array(ListTestimoniosResponseItem)


/**
 * @summary Compartir un testimonio
 */
export const CreateTestimonioBody = zod.object({
  "nombreAutor": zod.string(),
  "texto": zod.string()
})


/**
 * @summary Estadísticas del memorial
 */
export const GetStatsResponse = zod.object({
  "totalPersonas": zod.number(),
  "totalVelas": zod.number(),
  "totalRecuerdos": zod.number(),
  "totalTestimonios": zod.number(),
  "velasHoy": zod.number(),
  "recuerdosRecientes": zod.array(zod.object({
  "id": zod.number(),
  "personaId": zod.number().nullish(),
  "nombreAutor": zod.string(),
  "persona": zod.string().nullish(),
  "mensaje": zod.string(),
  "fotoUrl": zod.string().nullish(),
  "createdAt": zod.string(),
  "tiempoTranscurrido": zod.string()
})),
  "velasRecientes": zod.array(zod.object({
  "id": zod.number(),
  "personaId": zod.number().nullish(),
  "nombreRecordado": zod.string(),
  "nombreAutor": zod.string(),
  "mensaje": zod.string(),
  "createdAt": zod.string(),
  "tiempoTranscurrido": zod.string()
}))
})